[Makefile/build-sys] Adding a patching flow for rocket-chip & uncommenting SHA targets
1. Looking at this link from rocket-chip:
https://github.com/riscv/riscv-gnu-toolchain/blob/37fa6734566c78c54300dc3e0264d98e4e675bab/Makefile.in#L98
For some reason, piping curl into tar creates some problem with caching pipelines in Linux. On our server, this practice proved to be problematic with intermittent connectivity problems, which made debugging the flow harder. Also, I still don't understand why the flow wasn't designed to download a copy of the files and amortize on that rather than keep downloading them over-and-over again.
To fix this, I have created a patch that patches the build-system for riscv-tools. Particularly, I patched build.common to look for possible patches in a specified directory. If a patch exists for either configure.ac and/or Makefile.in for that particular tool, build.common would patch that and execute "autoconf".
There is, currently, one patch for riscv-gnu-tools. This patch would create a new target for the downloaded files and it would download the files normally.

2. This recently commented target would break the flow:
https://bitbucket.org/taylor-bsg/bsg_riscv/src/bcfadb45ff3ee648cf3d70a4bc103660a4eb0555/Makefile?at=master&fileviewer=file-view-default#Makefile-188:189
This target is needed by "build-sha", which is used to test integrating chisel-generated accelerators. As it is not included in rocket-chip repo, this breaks the flow. Therefore, I uncommented it back.